Reformulated directory structure and divided tests in files.
