+++ test/apollo-error-format.ts
@@ -1,0 +1,5 @@
+export interface ApolloErrorFormat {
+  message: string;
+  code: number;
+  details: string;
+}
